Squelch compiler warning

+++ src/UgrConnector.cc
@@ -357,7 +357,7 @@ void UgrConnector::do_n2n(std::string &path) {
   }
   
   // Make sure that no spurious queries enter Ugr
-  int pos = path.find('&');
+  size_t pos = path.find('&');
   if (pos != std::string::npos)
     path.erase(pos, std::string::npos);
